chore: fix build error
