+++ src/utils/cashout.utils.ts
@@ -1,20 +1,7 @@
-import crypto from 'crypto'
 import * as interfaces from '@/interfaces'
 import * as consts from '@/constants'
 import * as utils from '@/utils'
 import countries from 'i18n-iso-countries'
-export const hashPassword = (
-    password: string
-): {
-    salt: string
-    hash: string
-} => {
-    const salt = crypto.randomBytes(16).toString('hex')
-
-    const hash = crypto.pbkdf2Sync(password, salt, 10000, 64, 'sha512').toString('hex')
-
-    return { salt, hash }
-}
 
 export const convertPersonaUrl = (url: string) => {
     const parsedUrl = new URL(url)
